+++ simpleor/__init__.py
@@ -1,20 +1,12 @@
 """Top-level package for simpleor."""
-import sys
-import logging
+
 import simpleor.base
 import simpleor.scheduler
 import simpleor.cli
+from simpleor.monitoring import _configure_logger
 
 __author__ = """Lennart Damen"""
-__email__ = "lennart.damen@hotmail.com"
-__version__ = "0.0.4"
+__email__ = "lennart.damen.ai@gmail.com"
+__version__ = "0.0.5"
 
-
-logging.basicConfig(
-    stream=sys.stdout,
-    format="%(asctime)s | %(levelname)-8s | Process: %(process)d | %(name)s:%("
-    "funcName)s:%(lineno)d - %(message)s",
-    level=logging.DEBUG,
-)
-
-logger = logging.getLogger(f"{__name__}")
+_configure_logger(verbose="info")