+++ HISTORY.rst
@@ -2,7 +2,14 @@
 History
 =======
 
-0.0.0 (2020-10-25)
+0.0.4 (2020-12-12)
 ------------------
 
 * First release on PyPI.
+
+0.0.5 (2020-12-19)
+------------------
+
+* Added match maker module
+* Updated docs
+* Simplified GitHub workflow